Add files via upload
This code takes input from the user for the weight and the height. Then it calculates BMI and based on the result, the user is specified in a category (like: Underweight, Normal weight, ,Overweight, Obese) and offered some Health Recommendations. If BMI is lower than 16 or higher than 40, the program gives an alert.